+++ app/system/config.php
@@ -4,7 +4,7 @@ return [
 
     'application' => [
 
-        'version' => '0.11.0'
+        'version' => '0.11.1'
 
     ],