+++ app/system/modules/dashboard/app/components/widget-location.vue
@@ -5,6 +5,9 @@
             <li v-show="!editing">
                 <a class="pk-icon-contrast pk-icon-edit pk-icon-hover uk-hidden" :title="'Edit' | trans" data-uk-tooltip="{delay: 500}" @click.prevent="$parent.edit"></a>
             </li>
+            <li v-show="!editing">
+                <a class="pk-icon-contrast pk-icon-handle pk-icon-hover uk-hidden uk-sortable-handle" :title="'Drag' | trans" data-uk-tooltip="{delay: 500}"></a>
+            </li>
             <li v-show="editing">
                 <a class="pk-icon-delete pk-icon-hover" :title="'Delete' | trans" data-uk-tooltip="{delay: 500}" @click.prevent="$parent.remove" v-confirm="'Delete widget?'"></a>
             </li>