+++ app/system/modules/dashboard/app/components/widget-panel.vue
@@ -5,13 +5,16 @@
         <div class="uk-panel-badge" v-if="!type.disableToolbar">
             <ul class="uk-subnav pk-subnav-icon">
                 <li v-show="type.editable !== false && !editing">
-                    <a class="pk-icon-edit pk-icon-hover uk-hidden" title="{{ 'Edit' | trans }}" data-uk-tooltip="{delay: 500}" @click.prevent="edit"></a>
+                    <a class="pk-icon-edit pk-icon-hover uk-hidden" :title="'Edit' | trans" data-uk-tooltip="{delay: 500}" @click.prevent="edit"></a>
+                </li>
+                <li v-show="!editing">
+                    <a class="pk-icon-handle pk-icon-hover uk-hidden uk-sortable-handle" :title="'Drag' | trans" data-uk-tooltip="{delay: 500}"></a>
                 </li>
                 <li v-show="editing">
                     <a class="pk-icon-delete pk-icon-hover" :title="'Delete' | trans" data-uk-tooltip="{delay: 500}" @click.prevent="remove" v-confirm="'Delete widget?'"></a>
                 </li>
                 <li v-show="editing">
-                    <a class="pk-icon-check pk-icon-hover" title="{{ 'Close' | trans }}" data-uk-tooltip="{delay: 500}" @click.prevent="save"></a>
+                    <a class="pk-icon-check pk-icon-hover" :title="'Close' | trans" data-uk-tooltip="{delay: 500}" @click.prevent="save"></a>
                 </li>
             </ul>
         </div>