+++ app/system/modules/site/app/components/node-page.vue
@@ -1,7 +1,7 @@
 <template>
 
-    <div class="uk-grid pk-grid-large uk-form-stacked" data-uk-grid-margin>
-        <div class="uk-flex-item-1">
+    <div class="uk-grid pk-grid-large pk-width-sidebar-large uk-form-stacked" data-uk-grid-margin>
+        <div class="pk-width-content">
 
             <div class="uk-form-row">
                 <input class="uk-width-1-1 uk-form-large" type="text" name="page[title]" :placeholder="'Enter Title' | trans" v-model="page.title" v-validate:required lazy>
@@ -17,7 +17,7 @@
             </div>
 
         </div>
-        <div class="pk-width-sidebar pk-width-sidebar-large">
+        <div class="pk-width-sidebar">
 
             <div class="uk-panel">